Add support for fluent tagging, with nested tags.

+++ src/main/java/co/nstant/in/cbor/model/DataItem.java
@@ -16,7 +16,7 @@ public class DataItem {
         return majorType;
     }
 
-    public void setTag(int tag) {
+    public void setTag(long tag) {
         if (tag < 0) {
             throw new IllegalArgumentException("tag number must be 0 or greater");
         }
